smart-algo-trade/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
PyJWT==2.10.1
cryptography==41.0.7
requests==2.31.0
fyers-apiv3==3.1.7
websockets==12.0
aiohttp==3.9.3
python-multipart==0.0.6
numpy>=1.24.0
pandas>=2.0.0
pytz>=2023.3
ta-lib>=0.4.24
scikit-learn>=1.3.0
matplotlib>=3.8.0
scipy>=1.11.0
